+++ automaton/LeftParenAutomaton.cpp
@@ -1,6 +1,6 @@
 #include "LeftParenAutomaton.h"
 
-AutomatonResult LeftParenAutomaton::s0(const string &input, int currIndex, int currLine) {
+AutomatonResult LeftParenAutomaton::s0(const string &input, size_t currIndex, int currLine) {
     if (currIndex < input.length() && input[currIndex] == '(') {
         auto token = Token(TokenType::LEFT_PAREN, input.substr(currIndex, 1), currLine);
         return AutomatonSuccess(token, currIndex + 1, currLine);
