+++ automaton/RulesAutomaton.cpp
@@ -1,6 +1,6 @@
 #include "RulesAutomaton.h"
 
-AutomatonResult RulesAutomaton::s0(const string &input, int currIndex, int currLine) {
+AutomatonResult RulesAutomaton::s0(const string &input, size_t currIndex, int currLine) {
     if (currIndex + 5 <= input.length() && input.substr(currIndex, 5) == "Rules") {
         auto token = Token(TokenType::RULES, input.substr(currIndex, 5), currLine);
         return AutomatonSuccess(token, currIndex + 5, currLine);
