+++ automaton/UndefinedAutomaton.cpp
@@ -1,7 +1,7 @@
 #include <vector>
 #include "UndefinedAutomaton.h"
 
-AutomatonResult UndefinedAutomaton::s0(const string &input, int currIndex, int currLine) {
+AutomatonResult UndefinedAutomaton::s0(const string &input, size_t currIndex, int currLine) {
     if (currIndex < input.length()) {
         auto token = Token(TokenType::UNDEFINED, input.substr(currIndex, 1), currLine);
         return AutomatonSuccess(token, currIndex + 1, currLine);
